Fix rms_norm upcasting to float32 before Flash Attention
F.rms_norm computes in float32 internally; cast output back to
input dtype so Flash Attention receives bf16 as required.

+++ etude/gpt.py
@@ -49,7 +49,7 @@ class GPTConfig:
 
 
 def norm(x):
-    return F.rms_norm(x, (x.size(-1),))
+    return F.rms_norm(x, (x.size(-1),)).to(x.dtype)
 
 
 class Linear(nn.Linear):
